Refactor Scarpe module to class
