+++ lib/scarpe/version.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
-module Scarpe
+class Scarpe
   VERSION = "0.1.0"
 end
