+++ sig/scarpe.rbs
@@ -1,4 +1,4 @@
-module Scarpe
+class Scarpe
   VERSION: String
   # See the writing guide of rbs: https://github.com/ruby/rbs#guides
 end
